Replace limited spinner widget for selecting the payer.
The spinner widget is rather limited because it does not allow to have
no item selected by default (in order to force the user to make an
explicit choice), and it cannot show an error indicator. I'm replacing
it with a textview (using the spinner's style) and an alert dialog.